Fixing
What we did: Changed the commands for the lift in robot builder so that theres only one.  Used the shared pointer of the main controller to make another void statment in lift.cpp and .h files.

To Do: make movement methods in the new void statment using the shared joystick pointer to use buttons and axis joysticks